requirements.txt
alembic==1.3.0
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
autoenv==1.0.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==4.5.4
entrypoints==0.3
fakeredis==1.0.5
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
gunicorn==19.9.0
hiredis==1.0.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
locustio==0.13.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mirakuru==2.1.0
more-itertools==7.2.0
msgpack-python==0.5.6
packaging==19.2
pluggy==0.13.0
port-for==0.4
psutil==5.6.5
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.5
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
pyzmq==18.1.1
redis==3.3.11
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.10
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
